backend/requirements.txt
# Web server
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0

# Data processing and ML
numpy~=1.19.2
pandas==1.3.0
scikit-learn==0.24.2
xgboost==1.4.2
imbalanced-learn==0.8.0
joblib==1.0.1

# Image processing
pillow==8.3.1
tensorflow-cpu==2.5.0
opencv-python-headless==4.5.3.56

# Visualization and explanation
matplotlib==3.4.2
seaborn==0.11.1
